Clean up shell scripts
Removed parameter expansion formatting for shell variables. The old
formatting ${VAR} made the scripts a little harder to read.

Renamed the Polybar startup script from bin/polybar.sh to bin/polybar.
For some reason XMonad will not run the script unless its filename is
appended with .sh so the script is now sourced in xmonad.hs. Since
source is a bashism we will use POSIX compliant . because /bin/sh could
be linked to a POSIX shell.